move xgboost model builder together with xgboost model
