+++ pyzoo/zoo/automl/model/model_builder.py
@@ -14,10 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from abc import ABC, abstractmethod
-import logging
-
-logger = logging.getLogger(__name__)
+from abc import abstractmethod
 
 
 class ModelBuilder:
@@ -75,27 +72,3 @@ class PytorchModelBuilder(ModelBuilder):
         return model
 
 
-class XGBoostModelBuilder(ModelBuilder):
-
-    def __init__(self, model_type="regressor", cpus_per_trial=1, **xgb_configs):
-        self.model_type = model_type
-        self.model_config = xgb_configs.copy()
-
-        if 'n_jobs' in xgb_configs and xgb_configs['n_jobs'] != cpus_per_trial:
-            logger.warning(f"Found n_jobs={xgb_configs['n_jobs']} in xgb_configs. It will not take "
-                           f"effect since we assign cpus_per_trials(={cpus_per_trial}) to xgboost "
-                           f"n_jobs. Please raise an issue if you do need different values for "
-                           f"xgboost n_jobs and cpus_per_trials.")
-        self.model_config['n_jobs'] = cpus_per_trial
-
-    def build(self, config):
-        from zoo.orca.automl.xgboost.XGBoost import XGBoost
-        model = XGBoost(model_type=self.model_type, config=self.model_config)
-        model._build(**config)
-        return model
-
-    def build_from_ckpt(self, checkpoint_filename):
-        from zoo.orca.automl.xgboost.XGBoost import XGBoost
-        model = XGBoost(model_type=self.model_type, config=self.model_config)
-        model.restore(checkpoint_filename)
-        return model
